refactor(storage): extract notes_store + meetings_store from db.rs (gated, verbatim)
Pure MOVE, zero behavior change. The note-domain and meeting-domain
inherent impl Db methods lift out of storage/db.rs into two new files
storage/notes_store.rs (30 methods + 2 private row mappers) and
storage/meetings_store.rs (32 methods); db.rs 21,626 -> 20,445.

All 64 moved symbols are BYTE-IDENTICAL to trunk (verified incl.
doc-comments + attrs); the only non-relocation change is ONE forced
widening, MeetingStatus::as_str fn -> pub(crate) fn (the moved meeting
writers format the status column cross-file; mirrors the already
pub(crate) EntityKind::as_str), body untouched.

Gate/visibility/seal logic UNCHANGED, only relocated:
- 10 gated readers move verbatim, each still routing through
  visibility_clause: get_note_if_visible, note_is_visible,
  note_markdown_if_visible, list_notes_visible(_typed),
  latest_note_visible, list_meetings_visible, meeting_by_title(_folded)_visible,
  meeting_is_visible.
- the seal-on-write twins (insert_note_sealed, update_note_row_sealed,
  move_note_row_sealed, upsert_note_sealed, set_manual_notes_sealed,
  seal_manual_notes) relocate verbatim; their verify-before-destroy
  CALLERS stay in commands untouched.
Bias-to-move-less: delete_meeting, the whole search/index machinery,
list_link_candidates_visible, and seal/segment/timeline machinery stay
in db.rs.

Tests stay in db.rs mod tests (no test moved): db.rs #[test] 250 = trunk;
storage/ total 262 = trunk. Verified: cargo test --lib storage:: 262/0,
visible/sealed/lock/seal/gate/lifecycle/note_/meeting_/companion 696/0,
clippy --lib -D warnings clean (caught+fixed 3 now-dead db.rs imports),
MSRV grep clean.
